+++ Sources/X509/Verifier/Verifier.swift
@@ -13,147 +13,6 @@
 //===----------------------------------------------------------------------===//
 import SwiftASN1
 
-public struct VerificationDiagnostic: Sendable {
-    struct LeafCertificateHasUnhandledCriticalExtensions: Hashable {
-        var leafCertificate: Certificate
-        var handledCriticalExtensions: [ASN1ObjectIdentifier]
-    }
-    struct LeafCertificateIsInTheRootStoreButDoesNotMeetPolicy: Hashable {
-        var leafCertificate: Certificate
-        var failsToMeetPolicyReason: String
-    }
-    struct ChainFailsToMeetPolicy: Hashable {
-        var chain: UnverifiedCertificateChain
-        var failsToMeetPolicyReason: String
-    }
-    struct IssuerHasUnhandledCriticalExtension: Hashable {
-        var issuer: Certificate
-        var partialChain: [Certificate]
-        var handledCriticalExtensions: [ASN1ObjectIdentifier]
-    }
-    struct IssuerHasNotSignedCertificate: Hashable {
-        var issuer: Certificate
-        var partialChain: [Certificate]
-    }
-    enum Storage: Hashable {
-        case leafCertificateHasUnhandledCriticalExtension(LeafCertificateHasUnhandledCriticalExtensions)
-        case leafCertificateIsInTheRootStoreButDoesNotMeetPolicy(LeafCertificateIsInTheRootStoreButDoesNotMeetPolicy)
-        case chainFailsToMeetPolicy(ChainFailsToMeetPolicy)
-        case intermediateHashUnhandledCriticalExtension(IssuerHasUnhandledCriticalExtension)
-        case issuerHasNotSignedCertificate(IssuerHasNotSignedCertificate)
-    }
-    var storage: Storage
-}
-
-extension VerificationDiagnostic {
-    static func leafCertificateHasUnhandledCriticalExtension(
-        _ leafCertificate: Certificate,
-        handledCriticalExtensions: [ASN1ObjectIdentifier]
-    ) -> Self {
-        self.init(storage: .leafCertificateHasUnhandledCriticalExtension(
-            leafCertificate,
-            handledCriticalExtensions: handledCriticalExtensions
-        ))
-    }
-    static func leafCertificateIsInTheRootStoreButDoesNotMeetPolicy(
-        _ leafCertificate: Certificate,
-        reason failsToMeetPolicyReason: String
-    ) -> Self {
-        self.init(storage: .leafCertificateIsInTheRootStoreButDoesNotMeetPolicy(
-            leafCertificate,
-            reason: failsToMeetPolicyReason
-        ))
-    }
-    static func chainFailsToMeetPolicy(
-        _ chain: UnverifiedCertificateChain,
-        reason failsToMeetPolicyReason: String
-    ) -> Self {
-        self.init(storage: .chainFailsToMeetPolicy(
-            chain,
-            reason: failsToMeetPolicyReason
-        ))
-    }
-    static func intermediateHashUnhandledCriticalExtension(
-        issuer: Certificate,
-        chain: CandidatePartialChain,
-        handledCriticalExtensions: [ASN1ObjectIdentifier]
-    ) -> Self {
-        self.init(storage: .intermediateHashUnhandledCriticalExtension(
-            issuer: issuer,
-            partialChain: chain.chain + CollectionOfOne(chain.currentTip),
-            handledCriticalExtensions: handledCriticalExtensions
-        ))
-    }
-    static func issuerHasNotSignedCertificate(
-        _ issuer: Certificate,
-        chain: CandidatePartialChain
-    ) -> Self {
-        self.init(storage: .issuerHasNotSignedCertificate(
-            issuer,
-            partialChain: chain.chain + CollectionOfOne(chain.currentTip)
-        ))
-    }
-}
-
-extension VerificationDiagnostic.Storage {
-    static func leafCertificateHasUnhandledCriticalExtension(
-        _ leafCertificate: Certificate,
-        handledCriticalExtensions: [ASN1ObjectIdentifier]
-    ) -> Self {
-        .leafCertificateHasUnhandledCriticalExtension(.init(
-            leafCertificate: leafCertificate,
-            handledCriticalExtensions: handledCriticalExtensions
-        ))
-    }
-    static func leafCertificateIsInTheRootStoreButDoesNotMeetPolicy(
-        _ leafCertificate: Certificate,
-        reason failsToMeetPolicyReason: String
-    ) -> Self {
-        .leafCertificateIsInTheRootStoreButDoesNotMeetPolicy(.init(
-            leafCertificate: leafCertificate,
-            failsToMeetPolicyReason: failsToMeetPolicyReason
-        ))
-    }
-    static func chainFailsToMeetPolicy(
-        _ chain: UnverifiedCertificateChain,
-        reason failsToMeetPolicyReason: String
-    ) -> Self {
-        .chainFailsToMeetPolicy(.init(
-            chain: chain,
-            failsToMeetPolicyReason: failsToMeetPolicyReason
-        ))
-    }
-    static func intermediateHashUnhandledCriticalExtension(
-        issuer: Certificate,
-        partialChain: [Certificate],
-        handledCriticalExtensions: [ASN1ObjectIdentifier]
-    ) -> Self {
-        .intermediateHashUnhandledCriticalExtension(.init(
-            issuer: issuer,
-            partialChain: partialChain,
-            handledCriticalExtensions: handledCriticalExtensions
-        ))
-    }
-    static func issuerHasNotSignedCertificate(
-        _ issuer: Certificate,
-        partialChain: [Certificate]
-    ) -> Self {
-        .issuerHasNotSignedCertificate(.init(
-            issuer: issuer,
-            partialChain: partialChain
-        ))
-    }
-}
-
-extension VerificationDiagnostic: CustomStringConvertible {
-    /// Produces a human readable description of this ``VerificationDiagnostic`` that is potential expensive to compute.
-    ///
-    public var description: String {
-        // TODO: improve description for each kind
-        String(describing: storage)
-    }
-}
-
 public struct Verifier<Policy: VerifierPolicy> {
     public var rootCertificates: CertificateStore
 
@@ -248,7 +107,7 @@ public struct Verifier<Policy: VerifierPolicy> {
     private func shouldSkipAddingCertificate(partialChain: CandidatePartialChain, nextCertificate: Certificate, diagnosticCallback: ((VerificationDiagnostic) -> Void)?) -> Bool {
         // We want to confirm that the certificate has no unhandled critical extensions. If it does, we can't build the chain.
         if nextCertificate.hasUnhandledCriticalExtensions(handledExtensions: self.policy.verifyingCriticalExtensions) {
-            diagnosticCallback?(.intermediateHashUnhandledCriticalExtension(issuer: nextCertificate, chain: partialChain, handledCriticalExtensions: self.policy.verifyingCriticalExtensions))
+            diagnosticCallback?(.issuerHashUnhandledCriticalExtension(issuer: nextCertificate, chain: partialChain, handledCriticalExtensions: self.policy.verifyingCriticalExtensions))
             return true
         }
